+++ packages/app/sentry.client.config.ts
@@ -23,9 +23,6 @@ Sentry.init({
   // Enable performance monitoring through traces
   enableTracing: true,
 
-  // Track releases for better error monitoring
-  release: process.env.SENTRY_RELEASE || process.env.NEXT_PUBLIC_GITHUB_SHA || ENVIRONMENT,
-
   beforeSend(event) {
     // Filter out non-error events in production
     if (IS_PRODUCTION && !event.exception) return null;
